+++ panel/config.go
@@ -20,7 +20,7 @@ var (
 const (
 	// WDTT VPN-сервер
 	wdttConfigDir     = "/etc/wdtt"
-	wdttServerBin     = "/usr/local/bin/wdtt"
+	wdttServerBin     = "/usr/local/bin/wdtt-app"
 	wdttXrayRulesPath = "/usr/local/bin/wdtt-xray-rules.sh"
 	wdttServiceUnit   = "wdtt.service"